+++ btclib/hashes.py
@@ -22,7 +22,7 @@ from .utils import hash160, sha256
 
 def hash160_from_pubkey(pubkey: PubKey, compressed: Optional[bool] = None,
                         network: Optional[str] = None) -> Tuple[bytes, str]:
-    pubkey, network = bytes_from_pubkey(pubkey, compressed, network)
+    pubkey, network = bytes_from_pubkey(pubkey, network, compressed)
     h160 = hash160(pubkey)
     return h160, network